게시글 댓글 개수 관리 로직

+++ mvc-api/src/main/kotlin/com/jinuk/toy/mvcapi/view/post/response/PostDetailResponse.kt
@@ -13,7 +13,7 @@ data class PostDetailResponse(
     val content: String,
     val isViewerLike: Boolean,
     val likeCount: Int,
-    val commentCount: Int,
+    val commentCount: Long,
     val createdAt: LocalDateTime,
     val updatedAt: LocalDateTime,
 )
